feature: export internal hooks into API
