+++ change/@fluentui-react-combobox-b9579746-b4cf-441c-9538-35ad84994ace.json
@@ -1,0 +1,7 @@
+{
+  "type": "minor",
+  "comment": "feature: export internal hooks into API",
+  "packageName": "@fluentui/react-combobox",
+  "email": "bernardo.sunderhus@gmail.com",
+  "dependentChangeType": "patch"
+}
