+++ change/@fluentui-react-tag-picker-preview-af01b254-94fa-4050-b0f8-9b90fd4662f4.json
@@ -1,0 +1,7 @@
+{
+  "type": "patch",
+  "comment": "chore: adopt internal hooks provided by @fluentui/react-combobox",
+  "packageName": "@fluentui/react-tag-picker-preview",
+  "email": "bernardo.sunderhus@gmail.com",
+  "dependentChangeType": "patch"
+}
