+++ packages/react-components/react-combobox/src/components/Combobox/useInputTriggerSlot.ts
@@ -3,18 +3,19 @@ import type { ActiveDescendantImperativeRef } from '@fluentui/react-aria';
 import { mergeCallbacks, useEventCallback } from '@fluentui/react-utilities';
 import type { ExtractSlotProps, Slot, SlotComponentType } from '@fluentui/react-utilities';
 import { ArrowLeft, ArrowRight } from '@fluentui/keyboard-keys';
-import { useTriggerSlot, UseTriggerSlotState } from '../../utils/useTriggerSlot';
-import { ComboboxProps, ComboboxState } from './Combobox.types';
+import { ComboboxProps } from '../Combobox/Combobox.types';
+import { UseTriggerSlotState, useTriggerSlot } from '../../utils/useTriggerSlot';
+import { ComboboxBaseState } from '../../utils/ComboboxBase.types';
 import { OptionValue } from '../../utils/OptionCollection.types';
 import { getDropdownActionFromKey } from '../../utils/dropdownKeyActions';
 
 type UsedComboboxState = UseTriggerSlotState &
-  Pick<ComboboxState, 'value' | 'setValue' | 'selectedOptions' | 'clearSelection' | 'getOptionById'>;
+  Pick<ComboboxBaseState, 'value' | 'setValue' | 'selectedOptions' | 'clearSelection' | 'getOptionById'>;
 
 type UseInputTriggerSlotOptions = {
   state: UsedComboboxState;
   freeform: boolean | undefined;
-  defaultProps: Partial<ComboboxProps>;
+  defaultProps?: Partial<ComboboxProps>;
   activeDescendantController: ActiveDescendantImperativeRef;
 };
 
@@ -46,14 +47,14 @@ export function useInputTriggerSlot(
     activeDescendantController,
   } = options;
 
-  const onBlur = (ev: React.FocusEvent<HTMLInputElement>) => {
+  const onBlur = (event: React.FocusEvent<HTMLInputElement>) => {
     // handle selection and updating value if freeform is false
     if (!open && !freeform) {
       const activeOptionId = activeDescendantController.active();
       const activeOption = activeOptionId ? getOptionById(activeOptionId) : null;
       // select matching option, if the value fully matches
       if (value && activeOption && value.trim().toLowerCase() === activeOption?.text.toLowerCase()) {
-        selectOption(ev, activeOption);
+        selectOption(event, activeOption);
       }
 
       // reset typed value when the input loses focus while collapsed, unless freeform is true
@@ -84,8 +85,8 @@ export function useInputTriggerSlot(
   };
 
   // update value and active option based on input
-  const onChange = (ev: React.ChangeEvent<HTMLInputElement>) => {
-    const inputValue = ev.target.value;
+  const onChange = (event: React.ChangeEvent<HTMLInputElement>) => {
+    const inputValue = event.target.value;
     // update uncontrolled value
     setValue(inputValue);
 
@@ -94,7 +95,7 @@ export function useInputTriggerSlot(
 
     // clear selection for single-select if the input value no longer matches the selection
     if (!multiselect && selectedOptions.length === 1 && (inputValue.length < 1 || !matchingOption)) {
-      clearSelection(ev);
+      clearSelection(event);
     }
   };
 
@@ -120,27 +121,27 @@ export function useInputTriggerSlot(
    * Freeform combobox should not select
    */
   const defaultOnKeyDown = trigger.onKeyDown;
-  const onKeyDown = useEventCallback((ev: React.KeyboardEvent<HTMLInputElement>) => {
-    if (!open && getDropdownActionFromKey(ev) === 'Type') {
-      setOpen(ev, true);
+  const onKeyDown = useEventCallback((event: React.KeyboardEvent<HTMLInputElement>) => {
+    if (!open && getDropdownActionFromKey(event) === 'Type') {
+      setOpen(event, true);
     }
 
     // clear activedescendant when moving the text insertion cursor
-    if (ev.key === ArrowLeft || ev.key === ArrowRight) {
+    if (event.key === ArrowLeft || event.key === ArrowRight) {
       setHideActiveDescendant(true);
     } else {
       setHideActiveDescendant(false);
     }
 
     // update typing state to true if the user is typing
-    const action = getDropdownActionFromKey(ev, { open, multiselect });
+    const action = getDropdownActionFromKey(event, { open, multiselect });
     if (action === 'Type') {
       isTyping.current = true;
     }
     // otherwise, update the typing state to false if opening or navigating dropdown options
     // other actions, like closing the dropdown, should not impact typing state.
     else if (
-      (action === 'Open' && ev.key !== ' ') ||
+      (action === 'Open' && event.key !== ' ') ||
       action === 'Next' ||
       action === 'Previous' ||
       action === 'First' ||
@@ -152,12 +153,12 @@ export function useInputTriggerSlot(
     }
 
     // allow space to insert a character if freeform & the last action was typing, or if the popup is closed
-    if ((isTyping.current || !open) && ev.key === ' ') {
-      triggerFromProps?.onKeyDown?.(ev);
+    if ((isTyping.current || !open) && event.key === ' ') {
+      triggerFromProps?.onKeyDown?.(event);
       return;
     }
 
-    defaultOnKeyDown?.(ev);
+    defaultOnKeyDown?.(event);
   });
 
   trigger.onKeyDown = onKeyDown;
